Copy and paste the AdSense code in between the <head></head> tags of your site Place this code on every page across your site and Google will automatically show ads in all the best places for you
<script async src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-7031136340185694"
     crossorigin="anonymous"></script>
It can take up to an hour for ads to appear on the page. See our code implementation guide⁠‎ for more details.

place this script as said
NOTE:- that i have to show ads everywhere (on sidebar, on home page, on tools page, on each and every tools page too)

=====

<script async src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-7031136340185694"
     crossorigin="anonymous"></script>